add network caching
